refactor: move `XCUITestRunner.js` to `src/ios/`.
